NM modified files.

+++ config.py
@@ -68,7 +68,7 @@ config_defaults.update({
 'orthomosaic.depth_map'               : ['Medium','Aggressive'],
 'orthomosaic.epsg'                    : 32748,
 'orthomosaic.pixel_size'              : 0.025,
-'orthomosaic.output_type'             : 'Int16',
+'orthomosaic.output_type'             : 'Float32',
 'orthomosaic.python_path'             : python_path,
 'orthomosaic.scr_dir'                 : scr_dir,
 'orthomosaic.middle_left_frame_width' : 1000,
